requirements.txt
streamlit
numpy
gdown
pandas
scikit-surprise
